+++ prisma/mocks/plus.ts
@@ -1,7 +1,7 @@
 import { Prisma } from "@prisma/client";
-import plusStatusFactory from "../factories/plus_status";
-import plusSuggestionFactory from "../factories/plus_suggestion";
-import plusVotingSummaryFactory from "../factories/plus_voting_summary";
+import plusStatusFactory from "../factories/plusStatus";
+import plusSuggestionFactory from "../factories/plusSuggestion";
+import plusVotingSummaryFactory from "../factories/plusVotingSummary";
 
 // TODO: All this data is currently hard-coded so that Cypress tests will pass.
 // In the future, we could allow the data generated by these factories be more random, and
